Update to Debian 8.1, add missing packages, improve README.
Originally, the Freetz source was checked out into a shared folder.  At least
on my Mac, the build fails because the shared filesystem does not support
hard links.  Therefore, the checkout is now in a local directory.  The
images directory is symlinked to the host images directory, so that the firmware
image is available directly on the host.

The config file unfortunately cannot be handled in the same way, as
menuconfig will replace the symlink with a regular file once the config is
saved.  Instead, the bash profile and logout scripts are used to create a
symlink on login, and copy the file back on logout.  While this is not
perfect, it at least covers the common use case.